fix: refactor bytecode interpreter handling of call (#1848)
* feat: refactor arithv/shift and call

This refactors the handling of arithv and shift instructions to use
direct decoding.  It also removes most of the logical for handling
call/return, and this will then be reimplemented.


* fix: support improved call execution

This adds improved support for executing a call instruction by breaking
it into a pair of bytecodes, ENTER / LEAVE.  Furthermore, the target
address is embedded within the instruction itself, rather than being
obtained via a sequence of lookups.

Decoding for bitwise instructions was not distinguishing the different
cases, meaning that execution required two dispatches.

This also fixes an issue related to incorrect setting of the frame
pointer.

---------

+++ pkg/zkc/vm/internal/bytecode/util.go
@@ -91,7 +91,7 @@ func getRelativeOffset(offset uint32, target Address, width uint) uint32 {
 func packRegsIntoCodes(bytes []byte) []uint32 {
 	var (
 		nBytes = uint32(len(bytes))
-		ncodes = nCodesPackedSmall(nBytes)
+		ncodes = nCodesPackedSmall(uint(nBytes))
 		//
 		codes = make([]uint32, ncodes)
 	)
@@ -115,10 +115,10 @@ func packRegsIntoCodes(bytes []byte) []uint32 {
 	return codes
 }
 
-func nCodesPackedSmall(n uint32) uint32 {
+func nCodesPackedSmall(n uint) uint32 {
 	var (
 		// 4 bytes per code
-		ncodes = n / 4
+		ncodes = uint32(n) / 4
 	)
 	// Round up if necessary
 	if n%4 != 0 {
